feat(recipes): add CatalogSupplyId and RecipeSupplyId value objects with validation

+++ src/main/java/com/restock/platform/planning/domain/model/valueobjects/RecipeSupplyQuantity.java
@@ -3,7 +3,7 @@ package com.restock.platform.planning.domain.model.valueobjects;
 import jakarta.persistence.Embeddable;
 
 @Embeddable
-public record RecipeSupplyQuantity(Integer supplyQuantity) {
+public record RecipeSupplyQuantity(Double supplyQuantity) {
     public RecipeSupplyQuantity {
         if (supplyQuantity == null || supplyQuantity < 0) {
             throw new IllegalArgumentException("Supply quantity must not be null or negative");
@@ -11,10 +11,10 @@ public record RecipeSupplyQuantity(Integer supplyQuantity) {
     }
 
     public RecipeSupplyQuantity() {
-        this(0);
+        this(0.0);
     }
 
-    public Integer getSupplyQuantity() {
+    public Double getSupplyQuantity() {
         return supplyQuantity;
     }
 }
